Pull running jobs tracking out of engine meta
