+++ lib/oban/queue/basic_engine.ex
@@ -1,0 +1,141 @@
+defmodule Oban.Queue.BasicEngine do
+  @moduledoc false
+
+  @behaviour Oban.Queue.Engine
+
+  import Ecto.Query
+  import DateTime, only: [utc_now: 0]
+
+  alias Oban.{Config, Job, Repo}
+
+  @impl true
+  def init(%Config{} = conf, opts) do
+    meta =
+      opts
+      |> Keyword.put_new(:paused, false)
+      |> Keyword.put(:node, conf.node)
+      |> Keyword.put(:started_at, utc_now())
+      |> Keyword.put(:updated_at, utc_now())
+      |> Map.new()
+
+    {:ok, meta}
+  end
+
+  @impl true
+  def put_meta(_conf, %{} = meta, key, value) do
+    Map.put(meta, key, value)
+  end
+
+  @impl true
+  def refresh(_conf, %{} = meta) do
+    %{meta | updated_at: utc_now()}
+  end
+
+  @impl true
+  def fetch_jobs(_conf, %{paused: true}, _running), do: {:ok, []}
+
+  def fetch_jobs(_conf, %{limit: limit}, running) when map_size(running) >= limit do
+    {:ok, []}
+  end
+
+  def fetch_jobs(%Config{} = conf, %{} = meta, running) do
+    demand = meta.limit - map_size(running)
+
+    subset =
+      Job
+      |> select([:id])
+      |> where([j], j.state == "available")
+      |> where([j], j.queue == ^meta.queue)
+      |> order_by([j], asc: j.priority, asc: j.scheduled_at, asc: j.id)
+      |> limit(^demand)
+      |> lock("FOR UPDATE SKIP LOCKED")
+
+    updates = [
+      set: [state: "executing", attempted_at: utc_now(), attempted_by: [meta.node]],
+      inc: [attempt: 1]
+    ]
+
+    Repo.transaction(
+      conf,
+      fn ->
+        query =
+          Job
+          |> where([j], j.id in subquery(subset))
+          |> select([j, _], j)
+
+        case Repo.update_all(conf, query, updates) do
+          {0, nil} -> []
+          {_count, jobs} -> jobs
+        end
+      end
+    )
+  end
+
+  @impl true
+  def complete_job(%Config{} = conf, %Job{} = job) do
+    Repo.update_all(
+      conf,
+      where(Job, id: ^job.id),
+      set: [state: "completed", completed_at: utc_now()]
+    )
+
+    :ok
+  end
+
+  @impl true
+  def discard_job(%Config{} = conf, %Job{} = job) do
+    updates = [
+      set: [state: "discarded", discarded_at: utc_now()],
+      push: [
+        errors: %{attempt: job.attempt, at: utc_now(), error: format_blamed(job.unsaved_error)}
+      ]
+    ]
+
+    Repo.update_all(conf, where(Job, id: ^job.id), updates)
+
+    :ok
+  end
+
+  @impl true
+  def error_job(%Config{} = conf, %Job{} = job, seconds) when is_integer(seconds) do
+    %Job{attempt: attempt, id: id, max_attempts: max_attempts} = job
+
+    set =
+      if attempt >= max_attempts do
+        [state: "discarded", discarded_at: utc_now()]
+      else
+        [state: "retryable", scheduled_at: seconds_from_now(seconds)]
+      end
+
+    updates = [
+      set: set,
+      push: [errors: %{attempt: attempt, at: utc_now(), error: format_blamed(job.unsaved_error)}]
+    ]
+
+    Repo.update_all(conf, where(Job, id: ^id), updates)
+
+    :ok
+  end
+
+  @impl true
+  def snooze_job(%Config{} = conf, %Job{id: id}, seconds) when is_integer(seconds) do
+    updates = [
+      set: [state: "scheduled", scheduled_at: seconds_from_now(seconds)],
+      inc: [max_attempts: 1]
+    ]
+
+    Repo.update_all(conf, where(Job, id: ^id), updates)
+
+    :ok
+  end
+
+  # Helpers
+
+  defp seconds_from_now(seconds), do: DateTime.add(utc_now(), seconds, :second)
+
+  defp format_blamed(%{kind: kind, reason: error, stacktrace: stacktrace}) do
+    {blamed, stacktrace} = Exception.blame(kind, error, stacktrace)
+
+    Exception.format(kind, blamed, stacktrace)
+  end
+end
